renaming route -> handler & splitting search handlers into their own module
